Trick LastPass into recognising login screen
LastPass fails to recognise the login screen because we use the hidden attribute to hide the password screen while it is not needed.

With this fix, we show the data-hidden attribute first, and 1 second after page load, when LastPass should have done its check, we swap back to the hidden attribute.

Reason: hidden hides what we need to hide for everybody (including screenreader users and keyboard users), so that's what we really want to use (and effectively end up using 1 sec after page load).
Signed-off-by: Hidde de Vries <hidde@hiddedevries.nl>